+++ examples/Version4/CompassCar/CompassCar.ino
@@ -23,7 +23,7 @@ void setup() {
 
 void loop() {
   // Read compass heading
-  uint16_t heading = robot.compassRead();
+  int heading = robot.compassRead();
   int rotationSpeed = 100;  // Set rotation speed
   int compassDeadZone = 10; // +- 10° deadzone
